Working Version
Addressed issue regarding wrong or expired API key.

Then put an if statement regarding the parameters for sunrise and sunset. It's hacky but it provides junk data if no data for the location regarding sunrise and sunset are not found. If there is data then it will provide the real data.